Remove behaviour and spec
